pugs vote initialisation
as a test:
chat gpt was asked to recreate pugs vote (after giving it the entire pupsvotehandler file)
model used: o1 mini
additional instructions given: sql table structures, asked it to change the command structure (regarding /pugs add and remove). before it was asked to recreate and code entirely, i asked it to give me a summary of how it would make the changes to adapt the pugs role structure.